requirements.txt
streamlit
streamlit-calendar
lunarcalendar
PyGithub